+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+ - Deprecated `.into_vertex_buffer_any()` in favor of `.into()`.
  - Fixed values in uniform blocks being required to implement the `Default` trait.
  - Fixed buffer sometimes not always correctly unbinded.